requirements.txt
Babel==2.11.0
bidict==0.22.0
blinker==1.5
cachelib==0.9.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
eventlet==0.33.2
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login @ file:///Users/ktietz/demo/mc3/conda-bld/flask-login_1629875349535/work
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-Session==0.4.0
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_6adj7x0ejx/croot/jinja2_1666908137966/work
MarkupSafe @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_d4a9444f-bd4c-4043-b47d-cede33979b0fve7bm42r/croots/recipe/markupsafe_1654597878200/work
passlib==1.7.4
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
six==1.16.0
speaklater==1.3
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
